Created basic directories and other structures in order to organize our project since the beginning. Added first empty activities we will be working on.
