+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,22 @@
         android:theme="@style/Theme.MSPP_Project"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name="activities.SplashActivity"
+            android:exported="true">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+        </activity>
+        <activity
+            android:name="activities.RegisterActivity"
+            android:exported="false" />
+        <activity
+            android:name="activities.LoginActivity"
+            android:exported="false" />
+        <activity
+            android:name="activities.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
