+++ app/src/main/java/activities/LoginActivity.kt
@@ -1,0 +1,12 @@
+package activities
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import com.example.mspp_project.R
+
+class LoginActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_login)
+    }
+}
