+++ app/src/main/java/activities/SplashActivity.kt
@@ -1,0 +1,12 @@
+package activities
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import com.example.mspp_project.R
+
+class SplashActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_splash)
+    }
+}
